Fix code duplication for shutdown hook.
