+++ rocketmq-common/src/main/java/com/alibaba/rocketmq/common/ServerController.java
@@ -1,0 +1,7 @@
+package com.alibaba.rocketmq.common;
+
+public interface ServerController {
+
+    void shutdown();
+
+}
