+++ rocketmq-remoting/src/main/java/com/alibaba/rocketmq/remoting/netty/NettyServerConfig.java
@@ -153,6 +153,6 @@ public class NettyServerConfig implements Cloneable{
     
     @Override  
     public Object clone() throws CloneNotSupportedException {  
-        return (NettyServerConfig)super.clone();  
+        return super.clone();
     }  
 }
